fixed enter key not being prevented
